added user authentication

+++ visualisation/requests/alarm_delete.php
@@ -2,10 +2,11 @@
 
 	include('../data/mysql.php');
 	
-	$id = $_POST['id'];
-	
-	// add alarm to the mysql database
-	$query = "DELETE FROM alarms WHERE id=$id";
-	$result = mysql_query($query) or die('Error: ' . mysql_error());
-
+	if($_SESSION['userid']>0){
+		$id = $_POST['id'];
+		
+		// add alarm to the mysql database
+		$query = "DELETE FROM alarms WHERE id=$id";
+		$result = mysql_query($query) or die('Error: ' . mysql_error());
+	}
 ?>
